# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/products/{product_name}/projects:
        get:
            tags:
                - Project
            operationId: Project_ListProjects
            parameters:
                - name: product_name
                  in: path
                  description: To list the product name of the repository.
                  required: true
                  schema:
                    type: string
                - name: field_selector
                  in: query
                  description: Filter the list by field.
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.project.v1.ListsReply'
    /api/v1/products/{product_name}/projects/{project_name}:
        get:
            tags:
                - Project
            operationId: Project_GetProject
            parameters:
                - name: product_name
                  in: path
                  description: The name of the product the project belongs to.
                  required: true
                  schema:
                    type: string
                - name: project_name
                  in: path
                  description: The name of the project being retrieved.
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.project.v1.GetReply'
        post:
            tags:
                - Project
            operationId: Project_SaveProject
            parameters:
                - name: product_name
                  in: path
                  description: The name of the product the project belongs to.
                  required: true
                  schema:
                    type: string
                - name: project_name
                  in: path
                  description: The name of the project being created or updated.
                  required: true
                  schema:
                    type: string
                - name: insecureSkipCheck
                  in: query
                  description: Whether or not to skip validation.
                  schema:
                    type: boolean
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.project.v1.SaveRequest_Body'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.project.v1.SaveReply'
        delete:
            tags:
                - Project
            operationId: Project_DeleteProject
            parameters:
                - name: product_name
                  in: path
                  description: The name of the product the project belongs to.
                  required: true
                  schema:
                    type: string
                - name: project_name
                  in: path
                  description: The name of the project being deleted.
                  required: true
                  schema:
                    type: string
                - name: insecureSkipCheck
                  in: query
                  description: Whether or not to skip validation.
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.project.v1.DeleteReply'
components:
    schemas:
...
        api.project.v1.DeleteReply:
            type: object
            properties:
                message:
                    type: string
                    description: The message being returned.
            description: Defines the SaveReply message which is used to return a message after deleting a project.
        api.project.v1.GetReply:
            type: object
            properties:
                product:
                    type: string
                    description: The name of the product the project belongs to.
                name:
                    type: string
                    description: The name of the project.
                language:
                    type: string
                    description: The language used in the project.
            description: Defines the GetReply message which is used to return a specific project.
        api.project.v1.ListsReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.project.v1.GetReply'
                    description: The list of projects being returned.
            description: Defines the ListsReply message which is used to return a list of projects.
        api.project.v1.SaveReply:
            type: object
            properties:
                message:
                    type: string
                    description: The message being returned.
            description: Defines the SaveReply message which is used to return a message after creating or updating a project.
        api.project.v1.SaveRequest_Body:
            type: object
            properties:
                language:
                    type: string
                    description: The language used in the project.
            description: The request body for the project.
...